Create initial REST API documentation
The Resources document lists all currently known resources and will be converted
to a list of separate resource documents once the individual resource documents
are done. The Headers and Response Codes document contains a list of request and
response headers and the HTTP return codes. This is a work in progress and will
be expanded upon.

The service resource and the methods it supports are documented in the
Resource-Service.md document.

The server resource and the methods it supports are documented in the
Resource-Server.md document.